Removed player proxy
